Factor out the threadgate editor and add editing to post views
